duplicate EtherType in arp module as ProtocolType in order to avoid issue with recursion during parsing. update ArpPayload padding
